refactor: reorganize folder structure and make search component reusable
